+++ sandpack-client/src/clients/node/index.ts
@@ -1,5 +1,7 @@
 /* eslint-disable no-console,@typescript-eslint/no-explicit-any,prefer-rest-params,@typescript-eslint/explicit-module-boundary-types */
 
+import { Buffer } from "buffer";
+
 import { PREVIEW_LOADED_MESSAGE_TYPE, Nodebox } from "@codesandbox/nodebox";
 import type {
   FilesMap,
@@ -364,7 +366,10 @@ export class SandpackNode extends SandpackClient {
      */
     if (this.emulatorShellProcess?.state === "running") {
       Object.entries(modules).forEach(([key, value]) => {
-        if (Buffer.compare(value, this._modulesCache.get(key)) !== 0) {
+        if (
+          !this._modulesCache.get(key) ||
+          Buffer.compare(value, this._modulesCache.get(key)) !== 0
+        ) {
           this.emulator.fs.writeFile(key, value);
         }
       });